Rearrange project for easier testing
