+++ libdrivers/include/drivers/drivers.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Sergey Guralnik
+ * Copyright 2021 Sergey Guralnik
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -24,21 +24,10 @@
  * THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "driver.h"
+#pragma once
 
-#include <stddef.h>
+#include "txvc/driver.h"
 
-/* These symbols are defined in driver.ld */
-extern const struct txvc_driver __txvc_drivers_begin[];
-extern const struct txvc_driver __txvc_drivers_end[];
-
-const struct txvc_driver* txvc_enumerate_drivers(
-        bool (*fn)(const struct txvc_driver *d, const void *extra), const void *extra) {
-    for (const struct txvc_driver* d = __txvc_drivers_begin; d != __txvc_drivers_end; d++) {
-        if(!fn(d, extra)) {
-            return d;
-        }
-    }
-    return NULL;
-}
+extern const struct txvc_driver* txvc_enumerate_drivers(
+        bool (*fn)(const struct txvc_driver *d, const void *extra), const void *extra);
