+++ libtxvc/include/txvc/defs.h
