+++ libtxvc/include/txvc/driver.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Sergey Guralnik
+ * Copyright 2020 Sergey Guralnik
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -26,17 +26,28 @@
 
 #pragma once
 
-#include <stdint.h>
+#include "defs.h"
+
 #include <stdbool.h>
+#include <stdint.h>
 
-#include "txvc_defs.h"
+struct txvc_driver {
+    const char *name;
+    const char *help;
 
-extern void txvc_bit_vector_random(uint8_t* out, int outSz);
+    bool (*activate)(const char **argNames, const char **argValues);
+    bool (*deactivate)(void);
 
-extern bool txvc_bit_vector_equal(
-        const uint8_t* lhs, int lhsStart, int lhsEnd,
-        const uint8_t* rhs, int rhsStart, int rhsEnd);
+    int (*max_vector_bits)(void);
+    int (*set_tck_period)(int tckPeriodNs);
+    bool (*shift_bits)(int numBits,
+            const uint8_t *tmsVector,
+            const uint8_t *tdiVector,
+            uint8_t *tdoVector
+            );
+} TXVC_ALIGNED(16);
 
-extern int txvc_bit_vector_format_lsb(char* out, int outSz, const uint8_t* vector, int start, int end);
-extern int txvc_bit_vector_format_msb(char* out, int outSz, const uint8_t* vector, int start, int end);
+#define TXVC_DRIVER(name) \
+    static const struct txvc_driver txvc_driver_ ## name \
+        TXVC_SECTION(.txvc_driver) TXVC_USED TXVC_ALIGNED(16)
